# api/openapi.yaml
openapi: 3.1.0
info:
  title: Bit Learning API
  version: 1.0.0
  description: Authentication and user profile API for Bit Learning.
servers:
  - url: /
    description: Current server
tags:
  - name: System
  - name: Authentication
  - name: Users
paths:
  /healthz:
    get:
      operationId: healthCheck
      summary: Check service health
      tags: [System]
      responses:
        "200":
          description: The service is healthy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /v1/auth/register:
    post:
      operationId: register
      summary: Register a user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              email: user@example.com
              password: password123
              display_name: Bit Learner
      responses:
        "201":
          description: User registered successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenEnvelope"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "409":
          description: The email is already registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/auth/login:
    post:
      operationId: login
      summary: Log in
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: user@example.com
              password: password123
      responses:
        "200":
          description: Login succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenEnvelope"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          description: The email or password is incorrect.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/users/me:
    get:
      operationId: getCurrentUser
      summary: Get the current user profile
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      operationId: updateCurrentUser
      summary: Update the current user profile
      tags: [Users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
            example:
              display_name: New name
              bio: Learning Go
              avatar_url: https://example.com/avatar.png
      responses:
        "200":
          description: Updated user profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    InvalidRequest:
      description: The request body is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: One or more fields failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: A valid bearer token is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The requested user was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: An internal server error occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
        display_name:
          type: string
          maxLength: 100
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
    UpdateProfileRequest:
      type: object
      additionalProperties: false
      properties:
        display_name:
          type: string
          maxLength: 100
        bio:
          type: string
          maxLength: 1000
        avatar_url:
          type: string
          format: uri
          maxLength: 2048
    User:
      type: object
      required: [id, email, display_name, bio, avatar_url, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
        bio:
          type: string
        avatar_url:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Token:
      type: object
      required: [access_token, token_type, expires_at, user]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          const: Bearer
        expires_at:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
    HealthResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [status]
          properties:
            status:
              type: string
              const: ok
    UserEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/User"
    TokenEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Token"
    ErrorDetail:
      type: object
      required: [field, rule]
      properties:
        field:
          type: string
        rule:
          type: string
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"